requirements.txt
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
codecarbon==2.8.3
colorama==0.4.6
cryptography==44.0.1
Deprecated==1.2.18
distro==1.9.0
exceptiongroup==1.2.2
fief-client==0.20.0
Flask==3.1.0
Flask-Cors==5.0.0
gcloud==0.18.3
googleapis-common-protos==1.68.0
GPUtil==1.4.0
groq==0.18.0
h11==0.14.0
halo==0.0.31
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jwcrypto==1.5.6
log-symbols==0.0.14
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.0.2
nvidia-ml-py==12.570.86
oauth2client==4.1.3
pandas==2.2.3
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
PyGithub==2.6.0
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pynvml==12.0.0
pyparsing==3.2.1
Pyrebase4==4.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
pytz==2025.1
questionary==2.1.0
RapidFuzz==3.12.1
requests==2.29.0
requests-toolbelt==0.10.1
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
spinners==0.0.24
termcolor==2.3.0
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
yaspin==3.1.0
zipp==3.21.0
